Both conflicts were in the generated session-settings docs data, where
master's regeneration (window view settings and function_implementation
removed, allow_preliminary_distinct_abandoning and
query_plan_lower_array_join_function added) collided with this branch's
regeneration for the two new LIKE settings.

Resolved by taking master's generated data and re-adding
text_index_like_max_postings_rows_to_read and
use_text_index_like_pattern_bypass to it, with the settings explorer's
subtree counts recomputed for the two groups that gained an entry.

+++ src/Core/PostgreSQLProtocol.h
@@ -1330,7 +1330,7 @@ private:
     String value;
 
 public:
-    CommandComplete(Command cmd_, Int32 rows_count_)
+    CommandComplete(Command cmd_, UInt64 rows_count_)
     {
         value = enum_to_string[cmd_];
